do blogs the way svelte does blogs
more closely followed the code in the official svelte site. Grab date and slug from the filename instead of frontmatter. Do all the processing of markdown in _posts so that it exports a full list of all posts in ready to use json. use polka/send to make server responses easier and do some caching